Jesses changes

+++ app/controllers/wiki_controller.rb
@@ -36,6 +36,7 @@ class WikiController < ApplicationController
 
   def destroy
      @wiki = Wiki.find(params[:id])
+     authorize(@wiki)
 
      if @wiki.destroy
        flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
